nats msg splitting and merging for applier (mysql)
We limit nats msg size to 64M-4K, splitting a oversize msg to multiple segments. We use extra 4 bytes to save segment index (iSeg).

The segment index is required to identify a resent msg.

+++ drivers/mysql/driver.go
@@ -216,7 +216,7 @@ func (d *Driver) SetupNatsServer(logger hclog.Logger) (err error) {
 	nOpts := gnatsd.Options{
 		Host:       natsAddr.IP.String(),
 		Port:       natsAddr.Port,
-		MaxPayload: int32(g.NatsMaxPayload),
+		MaxPayload: g.NatsMaxPayload,
 		//HTTPPort:   8199,
 		LogFile: "/opt/log",
 		Debug:   true,
